+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+WIP v0.3.0 - <date>
+* Added config-powered cache service
 v0.2.0 - 2016-08-14 6:22am EST
 * Added Error handler
 * added logging